+++ jmx/jmx-api/src/main/java/uk/gov/justice/services/jmx/api/UnsupportedSystemCommandException.java
@@ -1,8 +1,0 @@
-package uk.gov.justice.services.jmx.api;
-
-public class UnsupportedSystemCommandException extends RuntimeException {
-
-    public UnsupportedSystemCommandException(final String message) {
-        super(message);
-    }
-}